move cli to src/rail/cli/rail
